requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.10
redis==4.6.0
passlib==1.7.4
PyJWT==2.9.0
pydantic==1.10.11
python-dotenv==1.0.0